requirements.txt
anyio==3.7.0
boto3==1.26.151
botocore==1.29.151
click==8.1.3
fastapi==0.97.0
h11==0.14.0
idna==3.4
jmespath==1.0.1
pydantic==1.10.9
pypng==0.20220715.0
python-dateutil==2.8.2
python-dotenv==1.0.0
qrcode==7.4.2
s3transfer==0.6.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.6.3
urllib3==1.26.16
uvicorn==0.22.0
